% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fulltransform.R
\name{align}
\alias{align}
\title{Align Mandible}
\usage{
align(sample, filename = "results", ext = "txt", colnames = TRUE)
}
\arguments{
\item{sample}{The input data frame. 3 columns (x, y, and z, in that order) with no titles.}

\item{filename}{The filename, without extension, in "quotation marks". Defaults to "results".}

\item{ext}{The extension of the file, in "quotation marks". Works with txt and xyz - not sure if it works with anything else. Defaults to "txt".}

\item{colnames}{If true, results file will have a header. Defaults to true.}
}
\description{
Does everything except import the initial data frame, right up to writing a results file.
Orients on xy, yz, and xz axes, and loops until they stay the same.
Also plots some lovely graphs so you can make sure it hasn't messed up.
}
\examples{
align(SK40, "myfile", "xyz", colnames = FALSE)
}

