% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/baseslice.R
\name{baseslice}
\alias{baseslice}
\title{Isolate the base}
\usage{
baseslice(sample, filename, folder)
}
\description{
Slices off the base (xmin + 10%), inverts it, and culls backfaces. Saves as a txt.
}

