% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/centre.R
\name{centre}
\alias{centre}
\title{Translate centrally}
\usage{
centre(sample)
}
\arguments{
\item{sample}{The input data. Three column data frame with lowercase headings "x", "y", and "z".}
}
\description{
Translate model so mid-x, mid-y, and min-z are on 0,0,0.
}
\examples{
centre()
}
\keyword{centre}

