% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cullbackfaces.R
\name{cull.backfaces}
\alias{cull.backfaces}
\title{Cull Backfaces}
\usage{
cull.backfaces()
}
\description{
Cull the backfaces of a point cloud based on vertex normals.
}
\examples{
cull.backfaces()
}

