% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edgelength.R
\name{edgelength}
\alias{edgelength}
\title{Finding the five longest edges}
\usage{
edgelength(x, y)
}
\description{
Finds and plots (on a pre-existing graph) the five longest edges of the convex hull.
}

