% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boundingbox.R
\name{getMinBBox}
\alias{getMinBBox}
\title{Find Minimum Bounding Box}
\usage{
getMinBBox(xy)
}
\arguments{
\item{xy}{The 2 column matrix.}
}
\description{
Find the minimum bounding box of a set of points defined by a 2 column matrix. Adapted from code by Daniel Wollschlaeger:
https://github.com/dwoll/RExRepos/blob/master/content/posts/diagBounding.md
}
\examples{
getMinBBox()
}
\keyword{bounding}

