% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wholeramus.R
\name{gonialarea}
\alias{gonialarea}
\title{Gonial Area Slice}
\usage{
gonialarea(sample, filename, folder)
}
\arguments{
\item{sample}{The input data frame. 3 named columns (x, y, and z, in that order).}
}
\description{
Isolates the gonial area, flips to horizontal, and culls backfaces (in a weird glitchy kind of way).
}

