% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fileimport.R
\name{importfiles}
\alias{importfiles}
\title{Import xyz and txt files}
\usage{
importfiles()
}
\description{
Imports xyz and txt files. Must be six columns: coordinates (x, y, z) and vertex normals (x, y, z).
}

