% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/longedge.R
\name{longedge}
\alias{longedge}
\title{Find longest edge of convex hull}
\usage{
longedge(x, y)
}
\arguments{
\item{x}{Values on the x axis.}

\item{y}{Values on the y axis.}
}
\description{
Finds the longest edge in the convex hull. Returns a data frame with the two points on the line.
}

