% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/leftrightflip.R
\name{lrflip}
\alias{lrflip}
\title{Left-Right Split & Flip}
\usage{
lrflip(sample, filename, folder)
}
\arguments{
\item{sample}{The input data frame. 3 named columns (x, y, and z, in that order).}
}
\description{
Splits the mandible in half sagittally, then rotates it so the buccal side is on top and culls backfaces.
}

