% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/leftrightsplit.R
\name{lrsplit}
\alias{lrsplit}
\title{Left-Right Split}
\usage{
lrsplit(sample, filename)
}
\arguments{
\item{sample}{The input data frame. 3 named columns (x, y, and z, in that order).}
}
\description{
Splits the mandible in half sagittally.
}

