% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/menemslice.R
\name{menemslice}
\alias{menemslice}
\title{Mental Eminence Slice}
\usage{
menemslice(sample, filename, folder, saveplots = TRUE, mandiblename)
}
\arguments{
\item{sample}{The input data frame. 3 named columns (x, y, and z, in that order).}
}
\description{
Isolates the mental eminence, flips to horizontal, and culls backfaces.
}

