% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/menemslicept1.R
\name{menemslicept1}
\alias{menemslicept1}
\title{Mental Eminence Slice Part 1}
\usage{
menemslicept1(sample, filename)
}
\arguments{
\item{sample}{The input data frame. 3 named columns (x, y, and z, in that order).}
}
\description{
Isolates the mental eminence, flips to horizontal, and culls backfaces. First part for calling from batch().
}

