% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/menemslicept2.R
\name{menemslicept2}
\alias{menemslicept2}
\title{Mental Eminence Slice Part 2}
\usage{
menemslicept2(sample, filename, folder)
}
\arguments{
\item{sample}{The input data frame. 3 named columns (x, y, and z, in that order).}
}
\description{
Isolates the mental eminence, flips to horizontal, and culls backfaces. Second part for calling from batch().
}

