% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chull.R
\name{pch}
\alias{pch}
\title{Plot convex hull}
\usage{
pch(x, y)
}
\arguments{
\item{x}{Values on the x axis.}

\item{y}{Values on the y axis.}
}
\description{
Plots a graph of the sample, then draws the convex hull around it.
}

