% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ramusslice.R
\name{ramusslice}
\alias{ramusslice}
\title{Posterior Ramus Slice}
\usage{
ramusslice(sample, filename, folder, saveplots = TRUE)
}
\arguments{
\item{sample}{The input data frame. 3 named columns (x, y, and z, in that order).}
}
\description{
Isolates the posterior ramus, flips to horizontal, and culls backfaces.
}

