% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/topfiveedges.R
\name{topfiveedges}
\alias{topfiveedges}
\title{Find the five longest edges of convex hull}
\usage{
topfiveedges(x, y)
}
\description{
Find the five longest edges of convex hull. Returns topfivep1 (one set of points), topfivep2 (the corresponding points), and topfive (topfivep1, then topfivep2) to parent environment.
}

