% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wholeramus.R
\name{wholeramus}
\alias{wholeramus}
\title{Ramus Slice}
\usage{
wholeramus(sample, filename, folder)
}
\arguments{
\item{sample}{The input data frame. 3 named columns (x, y, and z, in that order).}
}
\description{
Isolates the ramus, flips to horizontal, and culls backfaces.
}

