% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xytransform.R
\name{xytransform}
\alias{xytransform}
\title{Transforms on the XY axis}
\usage{
xytransform(sample)
}
\arguments{
\item{sample}{Input data frame. 3 columns, named "x", "y", and "z".}
}
\description{
Orients the mandible on the xy axis so the chin is pointing downwards.
}

