% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xztransform.R
\name{xztransform}
\alias{xztransform}
\title{Transforms on the XZ axis}
\usage{
xztransform(sample)
}
\arguments{
\item{sample}{Input data frame. 3 columns, named "x", "y", and "z".}
}
\description{
Orients the mandible on the xz axis.
}

