% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yzalign.R
\name{yzalign}
\alias{yzalign}
\title{Aligns the chin and jaw angle}
\usage{
yzalign(sample)
}
\arguments{
\item{sample}{Input data frame. 3 columns, named "x", "y", and "z".}
}
\description{
Orients the mandible so the chin and jaw angle are in alignment. To be used after yz transform and xy transform (in that order).
}

