% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yztransform.R
\name{yztransform}
\alias{yztransform}
\title{Transforms on the YZ axis}
\usage{
yztransform(sample)
}
\arguments{
\item{sample}{Input data frame. 3 columns, named "x", "y", and "z".}
}
\description{
Orients the mandible on the yz axis.
}

