% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fulltransformxyz.R
\name{align2}
\alias{align2}
\title{Align Mandible}
\usage{
align2(sample, filename, folder, slice = TRUE)
}
\arguments{
\item{sample}{The input data frame. 3 columns (x, y, and z, in that order).}
}
\description{
For calling from batch().
}

