% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/batch.R
\name{batch}
\alias{batch}
\title{Batch import and align}
\usage{
batch()
}
\description{
Calls align2 (and thus various other functions) in order to import and align a series of.xyz files. Files must be 3 columns (x, y, z) with no normals.
Will be updated later to include other functions/make it run a bit better.
No variables needed (see example).
}
\examples{
batch()
}

